requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pyodbc
pyngrok
jwt